aegis-platform/backend/app/services/google_calendar.py
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build

SCOPES = ['https://www.googleapis.com/auth/calendar.events']

def get_google_creds():
    # This function should handle the OAuth2 flow to get user credentials
    # For this MVP, we'll return None as a placeholder
    return None

def add_event_to_calendar(creds, summary, description, start_time):
    service = build('calendar', 'v3', credentials=creds)
    event = {
        'summary': summary,
        'description': description,
        'start': {
            'dateTime': start_time,
            'timeZone': 'America/Los_Angeles',
        },
        'end': {
            'dateTime': start_time, # Assuming 1-hour duration for simplicity
            'timeZone': 'America/Los_Angeles',
        },
    }
    event = service.events().insert(calendarId='primary', body=event).execute()
    return event
