aegis-platform/backend/requirements.txt
fastapi
python-dotenv
supabase
passlib[bcrypt]
python-jose[cryptography]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
uvicorn
python-decouple
